+++ src/app/post/[id]/update/page.tsx
@@ -13,14 +13,7 @@ import Spinner from '@/app/components/Spinner';
 import LayerSpinner from '@/app/components/Spinner/LayerSpinner';
 import useUpdatePost from '@/app/hooks/post/useUpdatePost';
 import useUpdatePostResource from '@/app/hooks/post/useUpdatePostResource';
-import { PostType } from '@/app/types/api/post';
-
-interface PostFormInput {
-  title: string;
-  type: PostType;
-  contents: string;
-  isOpen: '1' | '0';
-}
+import { PostFormInput } from '@/app/types/form/post';
 
 export default function UpdatePost() {
   const params = useParams<{ id: string }>();
