+++ src/app/post/create/page.tsx
@@ -10,14 +10,7 @@ import TitleInput from '@/app/components/Post/Form/TitleInput';
 import TypeSelect from '@/app/components/Post/Form/TypeSelect';
 import LayerSpinner from '@/app/components/Spinner/LayerSpinner';
 import useCreatePost from '@/app/hooks/post/useCreatePost';
-import { PostType } from '@/app/types/api/post';
-
-interface PostFormInput {
-  title: string;
-  type: PostType;
-  contents: string;
-  is_open: '1' | '0';
-}
+import { PostFormInput } from '@/app/types/form/post';
 
 export default function CreatePost() {
   const methods = useForm<PostFormInput>();
@@ -29,7 +22,7 @@ export default function CreatePost() {
       title: data.title,
       type: data.type,
       contents: data.contents,
-      is_open: data.is_open === '1',
+      is_open: data.isOpen === '1',
     });
   };
